Move adapter related code to a new file.
